some-fix

+++ app/controllers/items_controller.rb
@@ -34,7 +34,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if user_signed_in? && current_user.id == item.user_id
+      item.destroy
+    end
   end
 
   private
